Define illuminants with XYZ values directly
Converting from xyY results in XYZ values that round to the values used
in this commit, but using the un-rounded values causes a lot of tests to
fail due to the slightly different values. Most of the test values are
generated from libraries that use these XYZ illuminant values, so we're
using the same values they do rather than reducing the precision
requirements of our tests.